Move phase logic and movement into rugfatherAI from rugfather.js
